ArC: add Kotlin tests
