+++ independent-projects/arc/pom.xml
@@ -46,6 +46,7 @@
         <version.jta>2.0.1</version.jta>
         <version.jandex>3.0.5</version.jandex>
         <version.junit5>5.9.2</version.junit5>
+        <version.kotlin>1.8.10</version.kotlin>
         <version.maven>3.8.7</version.maven>
         <version.assertj>3.24.2</version.assertj>
         <version.jboss-logging>3.5.0.Final</version.jboss-logging>
